requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.6.15
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.2
cryptography==45.0.4
distlib==0.3.9
distro==1.9.0
filelock==3.18.0
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
identify==2.6.12
idna==3.10
iniconfig==2.1.0
jiter==0.10.0
lxml==5.4.0
multidict==6.5.0
nodeenv==1.9.1
openai==1.88.0
packaging==25.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
pre_commit==4.2.0
propcache==0.3.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.1
pypdfium2==4.30.1
pytest==8.4.0
pytest-asyncio==1.0.0
python-docx==1.2.0
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.4
slack_bolt==1.23.0
slack_sdk==3.35.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
virtualenv==20.31.2
yarl==1.20.1
